feat: 방 회전 기능을 FurnitureControllerBtn으로 이전
- CameraButtons.tsx의 기능을 FurnitureControllerBtn.tsx로 통합
